refatored minions to seperate route module
